make PanelMeasurement receive nodeSchema
